core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dEAdbEEF314", email = "deadbeef314@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.89",
    "prefect>=3.6.26",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
